AME-11980-implement-lru: this includes an LRU mechanism in cache.c, some signed/unsigned maths bugs in alloc.c and a slightly different way of getting the cache memory segments inside shm pools. The cache now uses environment setting to limit its size.
NOTE: the integrations tests no longer work because of the new dependency on utility.c. A fix for this is "in progress".

+++ source/share.c
@@ -37,7 +37,7 @@
  * the callback (cb) is called when the block is first opened and it is for initialisaiton of the block
  *
  */
-int get_memory_segment(am_shm_t **p_addr, char *name, size_t sz, void (*cb)(void *cbdata, void *p), void *cbdata, int id)
+int get_memory_segment(am_shm_t **p_addr, char *name, size_t sz, void (*cb)(void *cbdata, void *p), void *cbdata, unsigned long id)
 {
     if (p_addr == NULL) {
         return AM_FAIL;
